Tested the model, it failed so far.
Tried choosing stocks based on validation performance but this then failed when applied to test set, so this is a fail too.

Models with running mean of size p and slope thresholding are a nope.
This means that one cannot assume that if it was going up that it will continue to go up, even if that assumption is only for one timestep. This also means this occurs regardless of how many values you use to determine it's "going up", and it occurs regardless of at what point you consider the increase in value to be called "going up".

Or at least this is the case for 5-minute scales.

On to improved models! Next possible theories:

* The above, but with acceleration instead of velocity
* The above but with a minimum commitment period where it won't sell until after it's "held its cards" (probs not, i don't like this because it assumes that if it's gone up, then even if it goes down it will return if you wait)
* The above but with kinematics equation
* That increases and decreases happen at roughly constant velocity / acceleration for a given stock, so if it goes up at a speed, it will go down at the same speed. Meaning if we notice it going up, we invest, and keep that money until we notice it going down. - so that if the assumption holds then at worst case we will break mostly even, but best case we cash out on the beginning of the downhill after a large uphill increase.