Gonna make a new model and skip right to the blowtorch.
First off we're taking all the data for each timestep now - high, low, open, close, and volume. Then i'm gonna start throwing RL algorithms at it because each timestep can be treated as a state input, where its actions are buy/sell/nothing, it's reward is its net worth. Gonna work on that but for now i'm tired, and leaving it to make the new archive.