Implemented the optimization and that improved our previous frozen runtime of 326s / stock to 20s / stock !!!
That's about a 16x improvement, and from the original it's about 35x.

But then I realized wait, we're running it on all the samples at once now, let me try predict_on_batch! And that ran in fuckin 5 SECONDS PER STOCK. So now i'm working with that and putting this here before i delete the frozen model in favor of this new one!

Also going to see if i can improve it by changing batch size